+++ src/components/forms/components/index.js
@@ -1,3 +1,2 @@
 export { default as FormWithLinks } from './FormWithLinks'
-export { default as SemanticInput } from './SemanticInput'
 export { default as FormikInput } from './FormikInput'
